64-bit

+++ xlloption/xll_cumulant.cpp
@@ -239,7 +239,7 @@ public:
         using fms::sequence::power;
         using fms::sequence::factorial;
 
-        return sum(take(kappa.size(), &kappa[0]) * power(s,s) / factorial(1.));
+        return sum(take((long)kappa.size(), &kappa[0]) * power(s,s) / factorial(1.));
     }
 };
